Fix @alelom comments
